requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.19
langchain-google-genai>=0.0.6
langchain-voyageai>=0.1.0
voyageai>=0.1.3
llama-parse>=0.1.0
FlagEmbedding>=1.2.8
pydantic>=2.6.0
python-multipart>=0.0.7
